Create turnAroundTime + fixed some naming issues
I made a separate file after implementing some of the changes. I don't know if it was just me that the original file had some compiling errors. I fixed most of them. I made a new file cause I wasn't sure if it was just a problem on my end.